Move base64.js to src/utils; similarly for test-base64.js
