+++ build-system/dep-check-config.js
@@ -73,7 +73,7 @@ exports.rules = [
     filesMatching: 'ads/**/*.js',
     mustNotDependOn: 'src/**/*.js',
     whitelist: [
-      'ads/**->src/base64.js',
+      'ads/**->src/utils/base64.js',
       'ads/**->src/log.js',
       'ads/**->src/mode.js',
       'ads/**->src/url.js',
