+++ extensions/amp-ad-network-adsense-impl/0.1/test/test-amp-ad-network-adsense-impl.js
@@ -15,7 +15,7 @@
  */
 
 import {AmpAdNetworkAdsenseImpl} from '../amp-ad-network-adsense-impl';
-import {base64UrlDecodeToBytes} from '../../../../src/base64';
+import {base64UrlDecodeToBytes} from '../../../../src/utils/base64';
 import * as sinon from 'sinon';
 
 describe('amp-ad-network-adsense-impl', () => {
